Tests - HRRR - clean up and add missing files as links
The UofU version is looking for input files with the suffix 'apcp06'
when reading precipitation data. Adding these to the test suite as
symbolic links as a temporary solution until reading the proper 6th hour
forecast file is implemented.
Also removing un-needed f00 files as they shouldn't be read for
modeling.