HRRR data loading refactor to add support loading the sixth forecast hour
Major refactor on how wgrib HRRR data is loaded to support loading the
first and sixth forecast hour. This uses the introduced HrrrVariable
data class to take care of the mapping of HRRR to SMRF information. Also
removes the special casing for precipitation data and makes that a SMRF
configuration via the `hrrr_sixth_hour_variables` list.