WindNinja - Refactor edge interpolation
WindNinja 3.11 includes a bugfix where the topo domain is now properly
extracted. This caused the re-interpolation by SMRF onto the configured
topo resolution to have an empty (Nan) edge all around the domain, which
was previously only on the left and bottom. This updates the algorithm
to fill in missing edge values all around the domain.