HRRR - FileLoader - Remove loop over many dates.
The default use case to load HRRR data is for one timestep at a time and then
distribute the data before moving on to the next hour of a simulation. Hence,
the loop over a start and end date is removed for that input type, keeping the
support for other (CSV, NetCDF) alive.