Albedo - Rename decay start and end date config and break out method.
Rename the parameters to define a decay window and move the function to determine whether
the window is present to the distribute class for re-use. This prepares add a new post
fire based decay method that should also be triggered within the same time window.

The config parameters are now:
* start_decay (formerly date_method_start_decay)
* end_decay (formerly date_method_end_decay)