Distribute - Major overhaul of how data is initialized.
Unify the initialization of all distributed variables by using the methods from the base
class. This creates central attributes when variables need access to the topo or metadata
information. This cleans up a lot of code in distribution classes that were similar and is
now in the base class. Also renamed all classes to be more descriptive.

Only one TODO remains to improve how the precipitation is initialized.