Framework - DRY out distribution enqueuing logic
Add a method to the ImageData class that can be used to check whether an output variable
by a distribution class is requested. This DRY's out the logic for enqueuing in the model
framework entry point.

Note that 'precipitation' is the only class not following conventions for output variable
vs file name. Added a TODO (#32)